requirements.txt
asgiref==3.7.2
Django==5.0
image==1.5.33
Pillow==10.1.0
pypng==0.20220715.0
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
